+++ src/state/src/map/map.cpp
@@ -13,9 +13,7 @@ Map::Map(std::vector<std::vector<TerrainType>> map, size_t map_size)
 
 size_t Map::GetSize() const { return map_size; }
 
-TerrainType Map::GetTerrainType(int64_t x, int64_t y) const {
-    return map[x][y];
-}
+TerrainType Map::GetTerrainType(size_t x, size_t y) const { return map[x][y]; }
 
 TerrainType Map::GetTerrainType(double x, double y) const {
     return map[(int64_t)x][(int64_t)y];
